+++ CHANGELOG.rst
@@ -2,7 +2,7 @@ Development
 ======================
 * Fixed critical bug in :meth:`pulsar.net.HttpResponse.default_headers`
 * Added :meth:`pulsar.utils.http.Headers.pop` method.
-* **71 regression tests**
+* **81 regression tests**
 
 Version 0.2.1 - 2011-Dec-18
 =======================================